Extract b64 helper
