+++ components/header/web-push-subscription.js
@@ -1,19 +1,7 @@
 import React, {useState, useEffect} from 'react';
 import styled from 'styled-components';
 
-const base64ToUint8Array = (base64) => {
-    const padding = '='.repeat((4 - (base64.length % 4)) % 4);
-    const b64 = (base64 + padding).replaceAll('-', '+').replaceAll('_', '/');
-
-    const rawData = window.atob(b64);
-    const outputArray = new Uint8Array(rawData.length);
-
-    for (let i = 0; i < rawData.length; ++i) {
-        outputArray[i] = rawData.charCodeAt(i);
-    }
-
-    return outputArray;
-};
+import {base64ToUint8Array} from '../../utils/push-notification-helpers';
 
 const StyledSubscribeButton = styled.div`
     display: flex;
