+++ utils/push-notification-helpers.js
@@ -1,0 +1,13 @@
+export const base64ToUint8Array = (base64) => {
+    const padding = '='.repeat((4 - (base64.length % 4)) % 4);
+    const b64 = (base64 + padding).replaceAll('-', '+').replaceAll('_', '/');
+
+    const rawData = window.atob(b64);
+    const outputArray = new Uint8Array(rawData.length);
+
+    for (let i = 0; i < rawData.length; ++i) {
+        outputArray[i] = rawData.charCodeAt(i);
+    }
+
+    return outputArray;
+};
